refactor: fold resources.go into reconciler.go, rename BBR to IPP, consolidate defaults
Per Nir's review:
- Fold resources.go into reconciler.go for both externalmodel and
  externalprovider packages (reduces file fragmentation)
- Rename managed-by label values from bbr-* to ipp-* (BBR → IPP rename)
- Move shared defaults (LabelManagedBy, DefaultGatewayName,
  DefaultGatewayNamespace, DefaultRouteTimeout) to common/constants.go
- Remove break in MapProviderToModels to support future multi-provider refs

All unit tests pass. No runtime behavior change — label rename only
affects controller reconcilers which are wired in PR #184.
